Extract shared code from setup_travis and setup_circle_ci actions
