+++ fastlane/lib/fastlane/actions/setup_circle_ci.rb
@@ -1,13 +1,15 @@
+require 'fastlane/ci'
+
 module Fastlane
   module Actions
     class SetupCircleCiAction < Action
       def self.run(params)
-        unless should_run?(params)
+        unless Ci.should_run?(force: params[:force])
           UI.message("Not running on CI, skipping `setup_circle_ci`")
           return
         end
 
-        setup_keychain
+        Ci.setup_keychain
         setup_output_paths(params)
       end
 
@@ -22,34 +24,6 @@ module Fastlane
         ENV["GYM_OUTPUT_DIRECTORY"] = (root + "gym").to_s
         ENV["FL_BUILDLOG_PATH"] = (root + "buildlogs").to_s
         ENV["SCAN_INCLUDE_SIMULATOR_LOGS"] = true.to_s
-      end
-
-      def self.setup_keychain
-        unless ENV["MATCH_KEYCHAIN_NAME"].nil?
-          UI.message("Skipping Keychain setup as a keychain was already specified")
-          return
-        end
-
-        keychain_name = "fastlane_tmp_keychain"
-        ENV["MATCH_KEYCHAIN_NAME"] = keychain_name
-        ENV["MATCH_KEYCHAIN_PASSWORD"] = ""
-
-        UI.message("Creating temporary keychain: \"#{keychain_name}\".")
-        Actions::CreateKeychainAction.run(
-          name: keychain_name,
-          default_keychain: true,
-          unlock: true,
-          timeout: 3600,
-          lock_when_sleeps: true,
-          password: ""
-        )
-
-        UI.message("Enabling match readonly mode.")
-        ENV["MATCH_READONLY"] = true.to_s
-      end
-
-      def self.should_run?(params)
-        Helper.ci? || params[:force]
       end
 
       #####################################################
