+++ fastlane/lib/fastlane/actions/setup_travis.rb
@@ -1,34 +1,17 @@
+require 'fastlane/ci'
+
 module Fastlane
   module Actions
     class SetupTravisAction < Action
       def self.run(params)
         # Stop if not executed by CI
-        if !Helper.ci? && !params[:force]
+        unless Ci.should_run?(force: params[:force])
           UI.message("Currently not running on CI system, skipping travis setup")
           return
         end
 
         # Create a temporary keychain
-        password = "" # we don't need a password, as the keychain gets removed after each run anyway
-        keychain_name = "fastlane_tmp_keychain"
-        ENV["MATCH_KEYCHAIN_NAME"] = keychain_name
-        ENV["MATCH_KEYCHAIN_PASSWORD"] = password
-
-        UI.message("Creating temporary keychain: \"#{keychain_name}\".")
-        Actions::CreateKeychainAction.run(
-          name: keychain_name,
-          default_keychain: true,
-          unlock: true,
-          timeout: 3600,
-          lock_when_sleeps: true,
-          password: password
-        )
-
-        # Enable readonly mode for match by default
-        # we don't want to generate new identities and
-        # profiles on Travis usually
-        UI.message("Enabling readonly mode for Travis")
-        ENV["MATCH_READONLY"] = true.to_s
+        Ci.setup_keychain
       end
 
       #####################################################
